on my local machine : 

```
php test.php 
Compiling...
Testing...
2.719694s: php csvfind.php output.csv winkpad
0.709878s: php csvfind_v2.php output.csv winkpad
0.663501s: php -dopcache.enable_cli=1 -dopcache.jit_buffer_size=100M -dopcache.jit=1255 csvfind_v2.php output.csv winkpad
1.964410s: ./c-csvfind-test output.csv winkpad
4.233607s: ./rust-csvfind-test output.csv winkpad
1.690632s: ./rust-csvfindv2-test output.csv winkpad
```