@unconv 2023-06-04
@fezfez You deserve a medal, sir! It's actually the fastest version of them all. The JIT version is actually slower on my machine, on PHP 8.1

```console
$ php test.php
Compiling...
Testing...
0.960888s: php csvfind.php output.csv winkpad
0.396362s: php csvfind_v2.php output.csv winkpad
0.464657s: php -dopcache.enable_cli=1 -dopcache.jit_buffer_size=100M -dopcache.jit=1255 csvfind_v2.php output.csv winkpad
0.567590s: ./c-csvfind-test-orig output.csv winkpad
0.578368s: ./c-csvfind-test output.csv winkpad
1.219467s: ./rust-csvfind-test output.csv winkpad
0.487401s: ./rust-csvfindv2-test output.csv winkpad
```