@kocoten1992 2023-06-07
php `csvfind_v2.php` jit version on my machine was about 8-9% faster than non jit version (similar to fezfez number):

```
0.033569097518921 # jit
0.036731004714966
```